+++ controllers/categoriesController.js
@@ -15,71 +15,61 @@ const validateQuery = require("../validators/queryValidator");
 const categoriesController = {
   getCategories: asyncHandler(async (req, res) => {
     console.log("getCategories running from categoriesController...");
-    console.log("req.url:", req.url);
-    console.log("req.query:", req.query);
-    console.log("req.params:", req.params);
-    console.log("req.body:", req.body);
-
-    const { action, category } = req.query;
-    const msg = `${category} category ${
-      action === "edit" ? `updated` : `deleted`
-    }`;
+    const { action } = req.query;
+    const msg = `Category ${action === "edit" ? `updated` : `deleted`}`;
 
     res.render("categories", {
       title: "Categories",
       ...(action && { success: { msg } }),
     });
   }),
-  getCategoryItems: asyncHandler(async (req, res) => {
-    // Need to select items from category
-    console.log("getCategoryItems running...");
-    console.log("req.params:", req.params);
-    const { category } = req.params;
-    const items = await getItems(req.params);
-    const { action, upc } = req.query;
+  getCategoryItems: [
+    validateQuery,
+    asyncHandler(async (req, res) => {
+      // Need to select items from category
+      console.log("getCategoryItems running...");
+      const errors = validationResult(req);
+      errors.throw();
+      const { category } = req.params;
+      const items = await getItems(req.params);
+      const { action } = req.query;
 
-    // What if category does not exist
-    // Is a try...catch block needed?
-    // What if item is edited?
-    //  How to check what has changed
-    // What if item is deleted?
-    const msg = `${upc} item ${action === "edit" ? `updated` : `deleted`}`;
-    console.log("items:", items);
-    res.render("category", {
-      title: category,
-      category,
-      items,
-      ...(action && { success: { msg } }),
-    });
-  }),
+      // What if category does not exist
+      // Is a try...catch block needed?
+      // What if item is edited?
+      //  How to check what has changed
+      // What if item is deleted?
+      const msg = `Item ${action === "edit" ? `updated` : `deleted`}`;
+      res.render("category", {
+        title: category,
+        category,
+        items,
+        ...(action && { success: { msg } }),
+      });
+    }),
+  ],
   getAddCategory: [
     validateQuery,
-    asyncHandler(async (req, res, next) => {
+    asyncHandler(async (req, res) => {
       console.log("getAddCategory running...");
-      console.log("req.originalUrl:", req.originalUrl);
-      console.log("req.query:", req.query);
-      console.log("req.path:", req.path);
       // Is a try...catch block needed?
       // What if req.success is a value other than 'true'?
       // What if category is not in the database?
-      const errors = validationResult(req);
-      console.log("getAddCategory errors:", errors);
-      const { action, category } = req.query;
+      validationResult(req).throw();
+      const { action } = req.query;
       res.render("addCategory", {
         title: "Add Category",
         action: "add",
         path: "category/add",
         active: true,
-        ...(action && { success: { msg: `${category} category added` } }),
+        ...(action && { success: { msg: "Category added" } }),
       });
     }),
   ],
   getEditCategory: asyncHandler(async (req, res) => {
     console.log("getEditCategory running...");
-    console.log("req.url:", req.url);
-    console.log("req.params:", req.params);
     const { category } = req.params;
-    console.log(`edit/category/${category}`);
+
     res.render("editCategory", {
       title: "Edit Category",
       inputs: { category },
@@ -90,12 +80,9 @@ const categoriesController = {
   }),
   getDeleteCategory: asyncHandler(async (req, res) => {
     console.log("getDeleteCategory running...");
-    console.log("req.url:", req.url);
-    console.log("req.params:", req.params);
     const { category } = req.params;
     // Need to update all items impacted to 'unassigned' category
     // Need category id or category text
-    console.log("category:", category);
 
     res.render("deleteCategory", {
       title: "Delete Category",
@@ -109,8 +96,6 @@ const categoriesController = {
     validateCategory,
     asyncHandler(async (req, res) => {
       console.log("postAddCategory running...");
-      console.log("req.path:", req.path);
-      console.log("req.baseUrl:", req.baseUrl);
       const errors = validationResult(req);
 
       if (!errors.isEmpty()) {
@@ -127,11 +112,7 @@ const categoriesController = {
 
       const { category } = req.body;
       await insertCategory({ category });
-      // Redirect to the "add category" page
-      // Render successful message
-      res.redirect(`/category/add?action=add&category=${category}`);
-
-      // res.redirect(`/category/add`);
+      res.redirect("/category/add?action=add");
     }),
   ],
   postEditCategory: [
@@ -159,21 +140,16 @@ const categoriesController = {
       const { category: newCategory } = req.body;
       // Update existing category with new category
       await updateCategory({ prevCategory: category, newCategory });
-      // Redirect to the categories page
-      // Render successful message
-      res.redirect(`/categories?action=edit&category=${category}`);
+      res.redirect("/categories?action=edit");
     }),
   ],
   postDeleteCategory: [
     validatePassword,
     asyncHandler(async (req, res) => {
       console.log("postDeleteCategory running...");
-      console.log("req.url:", req.url);
-      console.log("req.query:", req.query);
-      console.log("req.params:", req.params);
-      console.log("req.body:", req.body);
       const errors = validationResult(req);
       const { category } = req.params;
+
       if (!errors.isEmpty()) {
         const localErrors = errors.mapped();
         console.log("localErrors:", localErrors);
@@ -189,11 +165,8 @@ const categoriesController = {
 
       // Need to delete category
       // Need to update all impacted items with new category
-      console.log("Right before deleteCategory()");
       await deleteCategory({ category }, updateItems);
-      // Redirect to the categories page
-      // Render successful message
-      res.redirect(`/categories?action=delete&category=${category}`);
+      res.redirect("/categories?action=delete");
     }),
   ],
 };