+++ controllers/itemsController.js
@@ -14,23 +14,18 @@ const validatePassword = require("../validators/passwordValidator");
 const itemsController = {
   getAddItem: asyncHandler(async (req, res) => {
     console.log("getAddItem running...");
-    const { added, upc } = req.query;
-
-    // How to output added item's info on success page?
+    const { action } = req.query;
 
     res.render("addItem", {
       title: "Add Item",
       action: "add",
       path: "item/add",
-      ...(added && { success: { msg: `${upc} added` } }),
+      ...(action && { success: { msg: "Item added" } }),
     });
   }),
   getEditItem: asyncHandler(async (req, res) => {
     console.log("getEditItem running...");
-    console.log("req.url:", req.url);
-    console.log("req.params:", req.params);
     const item = await getItem(req.params);
-    console.log(item);
 
     const { upc, category } = req.params;
     res.render("editItem", {
@@ -43,14 +38,9 @@ const itemsController = {
   }),
   getDeleteItem: asyncHandler(async (req, res) => {
     console.log("getDeleteItem running...");
-    console.log("req.url:", req.url);
-    console.log("req.params:", req.params);
     const item = await getItem(req.params);
-    console.log("item:", item);
-    // Need to populate the inputs
-    // Need item data
-    // Need category item is assigned to
     const { category, upc } = req.params;
+
     res.render("deleteItem", {
       title: "Delete Item",
       inputs: { ...item },
@@ -64,12 +54,10 @@ const itemsController = {
     asyncHandler(async (req, res) => {
       console.log("postAddItem running...");
       const errors = validationResult(req);
+
       if (!errors.isEmpty()) {
         const localErrors = errors.mapped();
 
-        console.log(errors);
-        console.log("localErrors");
-        console.log(localErrors);
         return res.status(400).render("addItem", {
           title: "Add Item",
           errors: { ...localErrors },
@@ -82,8 +70,6 @@ const itemsController = {
       // categoryExists returns a category object or undefined
       // Do I need a try...catch block here?
       const categoryExists = await getCategory(req.body);
-      console.log("categoryExists:", categoryExists);
-
       // If category does not exist
       //  Insert category
       // Update Item
@@ -94,7 +80,7 @@ const itemsController = {
       // Render successful message
       // res.redirect("/item/add");
       const { upc } = req.body;
-      res.redirect(`/item/add/?added=true&upc=${upc}`);
+      res.redirect("/item/add/?action=add");
     }),
   ],
   postEditItem: [
@@ -107,25 +93,6 @@ const itemsController = {
 
       if (!errors.isEmpty()) {
         const localErrors = errors.mapped();
-        console.log("errors:", errors);
-        console.log("localErrors:", localErrors);
-        /* const customValidationResults = validationResult.withDefaults({
-          formatter: (error) => {
-            console.log("Inside formatter");
-            console.log(error);
-            if (error)
-              return {
-                ...error,
-              };
-          },
-        });
-
-        const foo = customValidationResults(req);
-        const fooMapped = foo.mapped();
-        console.log(fooMapped); */
-
-        // console.log(errors);
-        // console.log("foo:", foo.mapped());
 
         return res.status(400).render("editItem", {
           title: "Edit Item",
@@ -147,17 +114,12 @@ const itemsController = {
       // Redirect to the item's category page
       // Render successful message
       // res.redirect(`/category/${req.body.category}`);
-      res.redirect(`/category/${req.body.category}/?edited=true&upc=${upc}`);
+      res.redirect(`/category/${req.body.category}/?action=edit`);
     }),
   ],
   postDeleteItem: [
     validatePassword,
     asyncHandler(async (req, res) => {
-      console.log("postDeleteCategory running...");
-      console.log("req.url:", req.url);
-      console.log("req.query:", req.query);
-      console.log("req.params:", req.params);
-      console.log("req.body:", req.body);
       const errors = validationResult(req);
       const { category, upc } = req.params;
       if (!errors.isEmpty()) {
@@ -177,7 +139,7 @@ const itemsController = {
       // Need to go back to category page
       // Render successful message
       // res.redirect(`/category/${category}`);
-      res.redirect(`/category/${req.body.category}/?deleted=true&upc=${upc}`);
+      res.redirect(`/category/${req.body.category}/?action=delete`);
     }),
   ],
 };